[utilities] Ensure slots in deque is always integer
With python3, the 'slots' calculation returns a decimal number which
deque() will not accept, thus throwing an exception. Always make sure
slots is an integer value.

+++ sos/utilities.py
@@ -209,7 +209,7 @@ class AsyncReader(threading.Thread):
         slots = None
         if sizelimit:
             sizelimit = sizelimit * 1048576  # convert to bytes
-            slots = sizelimit / self.chunksize
+            slots = int(sizelimit / self.chunksize)
         self.deque = deque(maxlen=slots)
         self.start()
         self.join()
